Updated do.py and the makefiles to sustain the new structure of examples and tests.
