+++ code-experiments/test/integration-test/Makefile_win_gcc.in
@@ -1,0 +1,42 @@
+## Makefile to build C example programs included with the COCO distribution
+## under Windows
+##
+## NOTE: We have tried to make this Makefile as generic and portable
+## as possible. However, we regularly test it using GNU make for Windows
+## only (actually the one from the gnuwin32 library only).
+##
+## To make the compilation work under Windows, please rename this file to
+## 'Makefile' after renaming the provided file with the same name to
+## something else or install Cygwin and run GNU make from within Cygwin
+## directly.
+
+LDFLAGS += -lm
+CCFLAGS = -g -ggdb -std=c89 -pedantic -Wall -Wextra -Wstrict-prototypes -Wshadow -Wno-sign-compare -Wconversion
+
+########################################################################
+## Toplevel targets
+all: test_coco test_instance_extraction test_biobj test_toy
+
+clean:
+	IF EXIST "coco.o" DEL /F coco.o
+	IF EXIST "test_coco.o" DEL /F test_coco.o 
+	IF EXIST "test_coco.exe" DEL /F test_coco.exe
+	IF EXIST "test_instance_extraction.o" DEL /F test_instance_extraction.o 
+	IF EXIST "test_instance_extraction.exe" DEL /F test_instance_extraction.exe
+
+########################################################################
+## Programs
+test_coco: test_coco.o coco.o
+	${CC} ${CCFLAGS} -o test_coco coco.o test_coco.o ${LDFLAGS}  
+	
+test_instance_extraction: test_instance_extraction.o coco.o
+	${CC} ${CCFLAGS} -o test_instance_extraction coco.o test_instance_extraction.o ${LDFLAGS}  
+
+########################################################################
+## Additional dependencies
+coco.o: coco.h coco.c
+	${CC} -c ${CCFLAGS} -o coco.o coco.c
+test_coco.o: coco.h coco.c test_coco.c
+	${CC} -c ${CCFLAGS} -o test_coco.o test_coco.c
+test_instance_extraction.o: coco.h coco.c test_instance_extraction.c
+	${CC} -c ${CCFLAGS} -o test_instance_extraction.o test_instance_extraction.c
