+++ code-experiments/test/integration-test/NMakefile
@@ -1,0 +1,36 @@
+##
+## Minimal nmake file to build the coco test suit and demo program.
+##
+## NOTE: Since Microsoft has not shipped a compliant C99 compiler, we
+## force the compiler to treat the source code as C++. We strive to
+## make this work, but it may be better to switch to a standards
+## compliant compiler such as GCC or the Intel C compiler.
+##
+LD = link.exe
+#CFLAGS = /nologo /Z7 /TP /Ox /D_CRT_SECURE_NO_WARNINGS /Dsnprintf=_snprintf
+CFLAGS = -g /nologo /Z7 /Ox /D_CRT_SECURE_NO_WARNINGS /Dsnprintf=_snprintf
+LDFLAGS = /nologo
+TARGETS = test_coco test_instance_extraction
+
+########################################################################
+## Toplevel targets
+all: $(TARGETS)
+
+clean:
+	del /F $(TARGETS) *.obj
+
+########################################################################
+## Programs
+
+test_coco.exe: test_coco.obj coco.obj
+	$(LD) $(LDFLAGS) -out:$@ $**
+
+test_instance_extraction.exe: test_instance_extraction.obj coco.obj
+	$(LD) $(LDFLAGS) -out:$@ $**
+
+########################################################################
+## Additional dependencies
+
+test_coco.o: coco.h
+
+test_instance_extraction.o: coco.h
