+++ do.py
@@ -5,13 +5,13 @@ from __future__ import division
 from __future__ import print_function
 
 import sys
-import re
 import os
 import shutil
 import tempfile
 import subprocess
 import platform
-from subprocess import call, check_output, STDOUT
+import time
+from subprocess import check_output, STDOUT
 
 ## Change to the root directory of repository and add our tools/
 ## subdirectory to system wide search path for modules.
@@ -19,8 +19,8 @@ os.chdir(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, os.path.abspath('code-experiments/tools'))
 
 from amalgamate import amalgamate
-from cocoutils import make, run, python, rscript
-from cocoutils import copy_file, copy_tree, expand_file, write_file
+from cocoutils import make, run, python
+from cocoutils import copy_file, expand_file, write_file
 from cocoutils import git_version, git_revision
 
 core_files = ['code-experiments/src/coco_suites.c',
@@ -42,18 +42,100 @@ def build_c():
     global release
     amalgamate(core_files + ['code-experiments/src/coco_runtime_c.c'],  'code-experiments/build/c/coco.c', release)
     copy_file('code-experiments/src/coco.h', 'code-experiments/build/c/coco.h')
-    copy_file('code-experiments/src/bbob2009_testcases.txt', 'code-experiments/build/c/bbob2009_testcases.txt')
+    copy_file('code-experiments/src/reference_values_hypervolume.txt', 'code-experiments/build/c/reference_values_hypervolume.txt')
     write_file(git_revision(), "code-experiments/build/c/REVISION")
     write_file(git_version(), "code-experiments/build/c/VERSION")
     make("code-experiments/build/c", "clean")
     make("code-experiments/build/c", "all")
 
-def test_c():
+def run_c():
     build_c()
     try:
-        make("code-experiments/build/c", "clean")
-        make("code-experiments/build/c", "all")
-        run('code-experiments/build/c', ['./coco_test', 'bbob2009_testcases.txt'])
+        run('code-experiments/build/c', ['./example_bbob2009'])
+        run('code-experiments/build/c', ['./example_biobj'])
+        run('code-experiments/build/c', ['./example_toy'])
+    except subprocess.CalledProcessError:
+        sys.exit(-1)
+
+def test_c():
+    build_c()
+    # Perform unit tests
+    build_c_unit_tests()
+    run_c_unit_tests()
+    # Perform integration tests
+    build_c_integration_tests()
+    run_c_integration_tests()
+    # Perform example tests
+    build_c_example_tests()
+    run_c_example_tests()
+        
+def build_c_unit_tests():
+    libraryPath = '';
+    fileName = ''
+    if ('win32' in sys.platform):
+        fileName = 'cmocka.dll'
+        if '64' in platform.machine():
+            libraryPath = 'code-experiments/test/unit-test/lib/win64'
+        elif ('32' in platform.machine()) or ('x86' in platform.machine()):
+            if 'cygwin' in os.environ['PATH']:
+                libraryPath = 'code-experiments/test/unit-test/lib/win32_cygwin'
+            else:
+                libraryPath = 'code-experiments/test/unit-test/lib/win32_mingw'
+    elif ('linux' in sys.platform):
+        libraryPath = 'code-experiments/test/unit-test/lib/linux'
+        fileName = 'libcmocka.so'
+    elif ('darwin' in sys.platform): #Mac
+        libraryPath = 'code-experiments/test/unit-test/lib/macosx'
+        fileName = 'libcmocka.dylib'
+        
+    if (len(libraryPath) > 0):
+        copy_file(os.path.join(libraryPath, fileName), 
+                  os.path.join('code-experiments/test/unit-test', fileName))        
+    copy_file('code-experiments/build/c/coco.c', 'code-experiments/test/unit-test/coco.c')
+    copy_file('code-experiments/src/coco.h', 'code-experiments/test/unit-test/coco.h')
+    make("code-experiments/test/unit-test", "clean")
+    make("code-experiments/test/unit-test", "all")
+
+def run_c_unit_tests():
+    try:
+        run('code-experiments/test/unit-test', ['./unit_test'])
+    except subprocess.CalledProcessError:
+        sys.exit(-1)
+
+def build_c_integration_tests():   
+    copy_file('code-experiments/build/c/coco.c', 'code-experiments/test/integration-test/coco.c')
+    copy_file('code-experiments/src/coco.h', 'code-experiments/test/integration-test/coco.h')
+    make("code-experiments/test/integration-test", "clean")
+    make("code-experiments/test/integration-test", "all")
+
+def run_c_integration_tests():
+    try:
+        run('code-experiments/test/integration-test', ['./test_coco', 'bbob2009_testcases.txt'])
+        run('code-experiments/test/integration-test', ['./test_instance_extraction'])
+    except subprocess.CalledProcessError:
+        sys.exit(-1)
+    
+def build_c_example_tests():
+    if os.path.exists('code-experiments/test/example-test'):
+        shutil.rmtree('code-experiments/test/example-test')
+        time.sleep(1) # Needed to avoid permission errors for os.makedirs
+    os.makedirs('code-experiments/test/example-test') 
+    copy_file('code-experiments/build/c/coco.c', 'code-experiments/test/example-test/coco.c')
+    copy_file('code-experiments/src/coco.h', 'code-experiments/test/example-test/coco.h')
+    copy_file('code-experiments/src/reference_values_hypervolume.txt', 'code-experiments/test/example-test/reference_values_hypervolume.txt')
+    copy_file('code-experiments/build/c/example_bbob2009.c', 'code-experiments/test/example-test/example_bbob2009.c')
+    copy_file('code-experiments/build/c/example_biobj.c', 'code-experiments/test/example-test/example_biobj.c')
+    copy_file('code-experiments/build/c/example_toy.c', 'code-experiments/test/example-test/example_toy.c')
+    copy_file('code-experiments/build/c/Makefile.in', 'code-experiments/test/example-test/Makefile.in')
+    copy_file('code-experiments/build/c/Makefile_win_gcc.in', 'code-experiments/test/example-test/Makefile_win_gcc.in')
+    make("code-experiments/test/example-test", "clean")
+    make("code-experiments/test/example-test", "all")
+        
+def run_c_example_tests():
+    try:
+        run('code-experiments/test/example-test', ['./example_bbob2009'])
+        run('code-experiments/test/example-test', ['./example_biobj'])
+        run('code-experiments/test/example-test', ['./example_toy'])
     except subprocess.CalledProcessError:
         sys.exit(-1)
 
@@ -67,23 +149,6 @@ def leak_check():
                     './coco_test', 'bbob2009_testcases.txt']
     run('code-experiments/build/c', valgrind_cmd)
     
-## C - multiobjective case
-def build_c_mo():
-    global release
-    amalgamate(core_files + ['code-experiments/src/coco_runtime_c.c'],  'code-experiments/build/c/mo/coco.c', release)
-    copy_file('code-experiments/src/coco.h', 'code-experiments/build/c/mo/coco.h')
-    write_file(git_revision(), "code-experiments/build/c/mo/REVISION")
-    write_file(git_version(), "code-experiments/build/c/mo/VERSION")
-    make("code-experiments/build/c/mo", "clean")
-    make("code-experiments/build/c/mo", "all")
-
-def test_c_mo():
-    build_c_mo()
-    try:
-        run('code-experiments/build/c/mo', ['./demo_mo', 'test'])
-    except subprocess.CalledProcessError:
-        sys.exit(-1)
-    
 ################################################################################
 ## Python 2
 def _prep_python():
@@ -91,7 +156,7 @@ def _prep_python():
     amalgamate(core_files + ['code-experiments/src/coco_runtime_c.c'],  'code-experiments/build/python/cython/coco.c', 
                release)
     copy_file('code-experiments/src/coco.h', 'code-experiments/build/python/cython/coco.h')
-    copy_file('code-experiments/src/bbob2009_testcases.txt', 'code-experiments/build/python/bbob2009_testcases.txt')
+    copy_file('code-experiments/src/reference_values_hypervolume.txt', 'code-experiments/build/python/reference_values_hypervolume.txt')
     expand_file('code-experiments/build/python/README.in', 'code-experiments/build/python/README',
                 {'COCO_VERSION': git_version()}) # hg_version()})
     expand_file('code-experiments/build/python/setup.py.in', 'code-experiments/build/python/setup.py',
@@ -178,24 +243,6 @@ def test_python3():
     os.environ.pop('PYTHON')
 
 ################################################################################
-## R
-def build_r():
-    global release
-    copy_tree('code-experiments/build/r/skel', 'code-experiments/build/r/pkg')
-    amalgamate(core_files + ['code-experiments/src/coco_runtime_r.c'],  'code-experiments/build/r/pkg/src/coco.c',
-               release)
-    copy_file('code-experiments/src/coco.h', 'code-experiments/build/r/pkg/src/coco.h')
-    expand_file('code-experiments/build/r/pkg/DESCRIPTION.in', 'code-experiments/build/r/pkg/DESCRIPTION',
-                {'COCO_VERSION': git_version()})  # FIXME: it seems that this doesn't work, because it writes '0.0.789' instead of 0.0.789
-    rscript('code-experiments/build/r/', ['code-experiments/tools/roxygenize'])
-    rusn('code-experiments/build/r', ['R', 'CMD', 'build', 'pkg'])
-
-def test_r():
-    build_r()
-    run('code-experiments/build/r', ['R', 'CMD', 'check', 'pkg'])
-    pass
-
-################################################################################
 ## Matlab
 def build_matlab():
     global release
@@ -203,7 +250,7 @@ def build_matlab():
     copy_file('code-experiments/src/coco.h', 'code-experiments/build/matlab/coco.h')
     write_file(git_revision(), "code-experiments/build/matlab/REVISION")
     write_file(git_version(), "code-experiments/build/matlab/VERSION")
-    run('code-experiments/build/matlab', ['matlab', '-nodisplay', '-nosplash', '-r', 'setup, exit'])
+    run('code-experiments/build/matlab', ['octave', '-nodisplay', '-nosplash', '-r', 'setup, exit'])
     
 ################################################################################
 ## Java
@@ -216,7 +263,7 @@ def build_java():
     run('code-experiments/build/java', ['javac', 'JNIinterface.java'])
     run('code-experiments/build/java', ['javah', 'JNIinterface'])
     
-    # Finds the path to the headers jni.h and jni_md.h under (platform-dependent)
+    # Finds the path to the headers jni.h and jni_md.h (platform-dependent)
     # and compiles the JNIinterface library (compiler-dependent). So far, only
     # the following cases are covered:
     
@@ -283,11 +330,8 @@ def test_java():
 def build():
     builders = [
         build_c,
-        build_c_mo,
-        build_matlab,
-        # build_octave, 
+        #build_matlab,
         build_python,
-        build_r,
         build_java, 
         build_examples
     ]
@@ -304,9 +348,7 @@ def build():
 
 def test():
     test_c()
-    test_c_mo()
     test_python()
-    test_r()
     test_java()
 
 def help():
@@ -316,26 +358,24 @@ Usage: do.py <command> <arguments>
 
 Available commands:
 
-  build          - Build C, Python and R modules
-  test           - Test C, Python and R modules
+  build          - Build C, Python and Java modules
+  test           - Test C, Python and Java modules
   build-c        - Build C framework
-  build-c-mo     - Build multiobjective C framework
   build-python   - Build Python modules
   build-python2  - Build Python 2 modules
   build-python3  - Build Python 3 modules
-  build-r        - Build R package
   build-matlab   - Build Matlab package
-  build-examples - Update examples to latest framework code
   build-java     - Build Java package
+  build-examples - Update examples to latest framework code
+  run-c          - Build and run examples from the C framework
   run-python     - Run a Python script with installed COCO module
                    Takes a single argument (name of Python script file)
-  test-c         - Run minimal test of C components
-  test-c-mo      - Run minimal test of multiobjective C components
-  test-python    - Run minimal test of Python module
-  test-python2   - Run minimal test of Python 2 module
-  test-python3   - Run minimal test of Python 3 module
-  test-r         - Run minimal test of R package
-  test-java      - Run minimal test of Java package
+  test-c         - Build and run unit tests, integration tests and examples
+                   from the C framework
+  test-python    - Build and run minimal test of Python module
+  test-python2   - Build and run  minimal test of Python 2 module
+  test-python3   - Build and run  minimal test of Python 3 module
+  test-java      - Build and run  minimal test of Java package
   leak-check     - Check for memory leaks
 
 
@@ -348,25 +388,22 @@ def main(args):
         help()
         sys.exit(0)
     cmd = args[0].replace('_', '-')
-    if cmd == 'build-c': build_c()
-    elif cmd == 'test-c': test_c()
-    elif cmd == 'build-c-mo': build_c_mo() 
-    elif cmd == 'test-c-mo': test_c_mo()   
-    elif cmd == 'build-python': build_python()
-    elif cmd == 'run-python': run_python(args[1])
-    elif cmd == 'test-python': test_python()
-    elif cmd == 'build-python2': build_python2()
-    elif cmd == 'test-python2': test_python2()
-    elif cmd == 'build-python3': build_python3()
-    elif cmd == 'test-python3': test_python3()
-    elif cmd == 'build-matlab': build_matlab()
-    elif cmd == 'build-r': build_r()
-    elif cmd == 'test-r': test_r()
-    elif cmd == 'build-java': build_java()
-    elif cmd == 'test-java': test_java()
-    elif cmd == 'build-examples': build_examples()
-    elif cmd == 'build': build()
+    if cmd == 'build': build()
     elif cmd == 'test': test()
+    elif cmd == 'build-c': build_c()
+    elif cmd == 'build-python': build_python()
+    elif cmd == 'build-python2': build_python2()
+    elif cmd == 'build-python3': build_python3()
+    elif cmd == 'build-matlab': build_matlab()
+    elif cmd == 'build-java': build_java()
+    elif cmd == 'build-examples': build_examples()
+    elif cmd == 'run-c': run_c()
+    elif cmd == 'run-python': run_python(args[1])
+    elif cmd == 'test-c': test_c()
+    elif cmd == 'test-python': test_python()
+    elif cmd == 'test-python2': test_python2()
+    elif cmd == 'test-python3': test_python3()
+    elif cmd == 'test-java': test_java()
     elif cmd == 'leak-check': leak_check()
     else: help()
